pyproject.toml
[project]
name = "todo-app"
version = "0.1.0"
description = "A console-based todo application with in-memory storage"
authors = [
    {name = "Khanzadi", email = "developer@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pytest>=7.0.0",
    "python-multipart>=0.0.20",
    "mcp>=1.0.0",  # Official MCP SDK for standardized protocol integration
    "fastapi>=0.127.0",
    "sqlmodel>=0.0.29",
    "psycopg2>=2.9.11",
    "passlib[bcrypt]>=1.7.4",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
]

[project.scripts]
start = "src.todo.cli:app"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "cli: Tests for CLI components",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["src/"]
omit = ["*/tests/*", "*/venv/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
